+++ NEWS.md
@@ -1,3 +1,12 @@
+# ClustImpute 0.2.0
+
+* It used to be the (strong) recommendation to center the data if a weight function is used (n_end >1). Now, by default, the scaling with the weight function is towards the global overall mean of each feature. Thus, for centered data there is almost no change (due the random imputation mechanism data with a true unknown mean of zero might have an empirical mean unequal to zero).
+* There is a check if the data is centered, and potentially a warning.
+* Added a plot function for ClustImpute results showing marginal distributions by cluster and feature. Type histogram and boxplot.
+* Added custom print function
+* Updated vignette showing new functionality
+* All changes above increase dependencies
+
 # ClustImpute 0.1.7
 
 * Removed dependency from psych package.